requirements.txt
requests==2.22.0
mysql-connector-python==8.0.18
Flask==1.1.1
Flask-Admin==1.5.5
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
PyMySQL==0.9.3
SQLAlchemy==1.3.12
Flask-RESTful==0.3.8
Flask-Cors==3.0.7
